pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pathao-api"
version = "1.0.0"
description = "Unofficial Pathao API client for Python"
readme = "README.md"
authors = [
    { name="Muktadirul Islam Mahi", email="muktadirul.05@email.com" }
]
license = {text = "MIT"}
dependencies = [
    "requests>=2.28.0",
    "python-dotenv>=1.0.0"
]
requires-python = ">=3.7"

[project.urls]
Homepage = "https://github.com/Muktadirul675/pathao-api"
